+++ src/BuildStage.groovy
@@ -1,6 +1,18 @@
 public class BuildStage implements Stage {
     private plugins = new StagePlugins()
     private decorations = new StageDecorations()
+    private String command = './bin/build.sh'
+    private String commandPrefix
+
+    public BuildStage withCommand(String command) {
+        this.command = command
+        return this
+    }
+
+    public BuildStage withCommandPrefix(String commandPrefix) {
+        this.commandPrefix = commandPrefix
+        return this
+    }
 
     @Override
     public Closure pipelineConfiguration() {
@@ -9,7 +21,7 @@ public class BuildStage implements Stage {
         return { ->
             decorations.apply() {
                 stage("build") {
-                    sh("./bin/build.sh")
+                    sh(getFullCommand())
                 }
             }
         }
@@ -17,5 +29,14 @@ public class BuildStage implements Stage {
 
     public void decorate(Closure decoration) {
         decorations.add(decoration)
+    }
+
+    public String getFullCommand() {
+        def result = command
+        if (commandPrefix) {
+            result = "${commandPrefix} ${result}"
+        }
+
+        return result
     }
 }