+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+- [Issue #103](https://github.com/manheim/jenkinsfile-pipeline/issues/103): BuildStage: Ability to customize build command
 - [Issue #39](https://github.com/manheim/jenkinsfile-pipeline/issues/39): Allow pipeline to run on Dockerfile
 - [Issue #44](https://github.com/manheim/jenkinsfile-pipeline/issues/44): ConfirmBeforeDeployPlugin: parameterize timeout
 - [Issue #98](https://github.com/manheim/jenkinsfile-pipeline/issues/98): Change DeployStage ./bin/deploy.sh to rely on $ENVIRONMENT, instead of argument